pyproject.toml
[project]
name = "flexify-analysis"
version = "0.1.0"
description = "Analysis of the data from my Flexify logs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
]
